requirements.txt
langchain_core
pydantic_settings
sqlalchemy
langgraph
langchain
langchain_groq
llama_parse
psycopg2
pandas
rapidfuzz
python-dotenv
watchdog==4.0.0
fastapi
python-multipart
uvicorn
langgraph-checkpoint-sqlite
aiosqlite
pwdlib[argon2]
email-validator
click
tenacity
pyjwt
requests
sseclient
langchain-google-genai